dmm-passport middleware is no longer required
