+++ src/routing/game.js
@@ -8,36 +8,35 @@ const kancolle = require('../kancolle/')
 const URL = require('url-parse')
 const urljoin = require('url-join')
 const log = require('../logger')('app:router')
-const dmmPassport = require('../middleware/dmm-passport')
 const Cookie = require('tough-cookie').Cookie
+const CookieInjector = require('../dmm/cookie-injector')
 
 router.get('/', (req, res) => {
   res.render('index')
 })
 
-router.post('/dmm-account', passport.authenticate('local', {
+router.post('/dmm-account', passport.authenticate('dmm-account', {
   failureRedirect: '/',
   session: false
 }), (req, res, next) => {
   const dmmCookies = req.user.map(Cookie.parse)
   req.body.dmm_session = dmmCookies.find(cookie => cookie.key === 'INT_SESID').value
   next()
-}, loginByDmmSession)
+}, launchKancolle)
 
-router.post('/dmm-session', loginByDmmSession)
+router.post('/dmm-session', launchKancolle)
 
-function loginByDmmSession(req, res, next) {
-  dmmPassport.serialize([`INT_SESID=${req.body.dmm_session}`], (error, cookies) => {
-    log.info(`OSAPI: get DMM game metadata of app id ${kancolle.appId}`)
-    osapi.getGameInfo(kancolle.appId, cookies)
-    .then(kancolle.launch)
-    .then(redirectKancolleNetworkTraffic)
-    .then(url => {
-      log.info('render HTML template "kancolle" with Kancolle game url')
-      return res.render('kancolle', {flashUrl: url})
-    })
-    .catch(next)
+function launchKancolle(req, res, next) {
+  const cookies = injectCookies(`INT_SESID=${req.body.dmm_session}`, ['/', '/netgame/', '/netgame_s/'])
+  log.info(`OSAPI: get DMM game metadata of app id ${kancolle.appId}`)
+  osapi.getGameInfo(kancolle.appId, cookies)
+  .then(kancolle.launch)
+  .then(redirectKancolleNetworkTraffic)
+  .then(url => {
+    log.info('render HTML template "kancolle" with Kancolle game url')
+    return res.render('kancolle', {flashUrl: url})
   })
+  .catch(next)
 }
 
 function redirectKancolleNetworkTraffic(url) {
@@ -61,6 +60,15 @@ function redirectKancolleNetworkTraffic(url) {
     `?api_starttime=${url.query.api_starttime}`)
 
   return Promise.resolve(interceptedUrl)
+}
+
+function injectCookies(session, subdomains) {
+  const injector = new CookieInjector([session], subdomains)
+  log.verbose('set DMM to display language in Japanese')
+  injector.language(CookieInjector.language.japan)
+  log.verbose('revoke region restriction')
+  injector.revokeRegionRestriction()
+  return injector.cookies
 }
 
 module.exports = router
